+++ R/demo_simple_gmm.R
@@ -20,9 +20,9 @@ demo_simple_gaussian <- function(truemu=5, truesd=1, noisemu=0, noisesd=2, n=100
       noisydata <- data + noise
 
       truemodel <- gmm(nComp=1, mu=truemu, sigma=(truesd^2), weights=1, d=1)
-      score_q = pryr::partial(scorefunctiongmm, model=truemodel)
-      originalresult <- KSD(data,score_q=score_q,'rbf',-1.0)
-      noisyresult <- KSD(noisydata,score_q=score_q,'rbf',-1.0)
+      score_function = pryr::partial(scorefunctiongmm, model=truemodel)
+      originalresult <- KSD(data,score_function=score_function,'rbf',-1.0)
+      noisyresult <- KSD(noisydata,score_function=score_function,'rbf',-1.0)
 
       print(sprintf('p for fitting true distribution with original data: %f',originalresult$p))
       print(sprintf('p for fitting true distribution with data with noise: %f',noisyresult$p))
@@ -39,14 +39,14 @@ demo_normal_performance <- function(){
       n=100; truemu=0; truesd=1;noisemu=0;
       data <- rnorm(n, mean=truemu, sd=truesd)
       model <- gmm(nComp=1, mu=truemu, sigma=truesd, weights=1, d=1)
-      score_q = pryr::partial(scorefunctiongmm, model=model)
+      score_function = pryr::partial(scorefunctiongmm, model=model)
       #noisemu=seq(0,1,0.01);noisesd=rep(1,length(noisemu)); noiseaxis = noisemu;
       noisesd=seq(0,1,0.01);noisemu=rep(0,length(noisesd)); noiseaxis = noisesd
       pval = rep(0,length(noisesd))
       for (i in 1:length(noisesd)){
             noise <- rnorm(n,mean=noisemu[i],sd=noisesd[i])
             noisydata <- data + noise
-            result <- KSD(noisydata,score_q=score_q,'rbf',-1.0)
+            result <- KSD(noisydata,score_function=score_function,'rbf',-1.0)
             pval[i] = result$p
       }
 
@@ -75,9 +75,9 @@ demo_simple_gamma <- function(trueshape=10, truescale=3, noisemu=5, noisesd=2, n
       noise <- rnorm(n,mean=noisemu,sd=noisesd)
       noisydata <- data + noise
 
-      score_q = pryr::partial(gamma_score, shape=trueshape, scale=truescale)
-      originalresult <- KSD(data,score_q=score_q,'rbf',-1.0)
-      noisyresult <- KSD(noisydata,score_q=score_q,'rbf',-1.0)
+      score_function = pryr::partial(gamma_score, shape=trueshape, scale=truescale)
+      originalresult <- KSD(data,score_function=score_function,'rbf',-1.0)
+      noisyresult <- KSD(noisydata,score_function=score_function,'rbf',-1.0)
 
       print(sprintf('p for fitting true distribution with original data: %f',originalresult$p))
       print(sprintf('p for fitting true distribution with data with noise: %f',noisyresult$p))
@@ -92,48 +92,48 @@ demo_gmm <- function(){
       print('1. 1-dimensional Gaussian, 100 samples')
       model <- gmm()
       X <- rgmm(model, n=100)
-      score_q = pryr::partial(scorefunctiongmm, model=model)
-      result <- KSD(X,score_q=score_q, 'rbf',-1.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model)
+      result <- KSD(X,score_function=score_function, 'rbf',-1.0)
 
       model2 <- perturbgmm(model)
-      score_q = pryr::partial(scorefunctiongmm, model=model2)
-      result2 <- KSD(X,score_q=score_q, 'rbf',-1.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model2)
+      result2 <- KSD(X,score_function=score_function, 'rbf',-1.0)
       print(sprintf('Original p : %f',result$p))
       print(sprintf('Original p : %f',result2$p))
 
       print('2. 1-dimensional Gaussian, 300 samples')
       model <- gmm()
       X <- rgmm(model, n=300)
-      score_q = pryr::partial(scorefunctiongmm, model=model)
-      result <- KSD(X,score_q=score_q, 'rbf',-1.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model)
+      result <- KSD(X,score_function=score_function, 'rbf',-1.0)
 
       model2 <- perturbgmm(model)
-      score_q = pryr::partial(scorefunctiongmm, model=model2)
-      result2 <- KSD(X,score_q=score_q, 'rbf',-1.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model2)
+      result2 <- KSD(X,score_function=score_function, 'rbf',-1.0)
       print(sprintf('Original p : %f',result$p))
       print(sprintf('Original p : %f',result2$p))
 
       print('3. 1-dimensional Gaussian, 100 samples, median-heuristic')
       model <- gmm()
       X <- rgmm(model, n=100)
-      score_q = pryr::partial(scorefunctiongmm, model=model)
-      result <- KSD(X,score_q=score_q, 'rbf',-2.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model)
+      result <- KSD(X,score_function=score_function, 'rbf',-2.0)
 
       model2 <- perturbgmm(model)
-      score_q = pryr::partial(scorefunctiongmm, model=model2)
-      result2 <- KSD(X,score_q=score_q, 'rbf',-2.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model2)
+      result2 <- KSD(X,score_function=score_function, 'rbf',-2.0)
       print(sprintf('Original p : %f',result$p))
       print(sprintf('Original p : %f',result2$p))
 
       print('4. 1-dimensional Gaussian, 300 samples, median-heuristic')
       model <- gmm()
       X <- rgmm(model, n=300)
-      score_q = pryr::partial(scorefunctiongmm, model=model)
-      result <- KSD(X,score_q=score_q, 'rbf',-2.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model)
+      result <- KSD(X,score_function=score_function, 'rbf',-2.0)
 
       model2 <- perturbgmm(model)
-      score_q = pryr::partial(scorefunctiongmm, model=model2)
-      result2 <- KSD(X,score_q=score_q, 'rbf',-2.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model2)
+      result2 <- KSD(X,score_function=score_function, 'rbf',-2.0)
       print(sprintf('Original p : %f',result$p))
       print(sprintf('Original p : %f',result2$p))
 }
@@ -145,48 +145,48 @@ demo_gmm_multi <- function(){
       print('1. 3-dimensional Gaussian, 100 samples')
       model <- gmm(nComp=5,d=3)
       X <- rgmm(model, n=100)
-      score_q = pryr::partial(scorefunctiongmm, model=model)
-      result <- KSD(X,score_q=score_q, 'rbf',-1.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model)
+      result <- KSD(X,score_function=score_function, 'rbf',-1.0)
 
       model2 <- perturbgmm(model)
-      score_q = pryr::partial(scorefunctiongmm, model=model2)
-      result2 <- KSD(X,score_q=score_q, 'rbf',-1.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model2)
+      result2 <- KSD(X,score_function=score_function, 'rbf',-1.0)
       print(sprintf('Original p : %f',result$p))
       print(sprintf('Original p : %f',result2$p))
 
       print('2. 3-dimensional Gaussian, 300 samples')
       model <- gmm(nComp=5,d=3)
       X <- rgmm(model, n=300)
-      score_q = pryr::partial(scorefunctiongmm, model=model)
-      result <- KSD(X,score_q=score_q, 'rbf',-1.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model)
+      result <- KSD(X,score_function=score_function, 'rbf',-1.0)
 
       model2 <- perturbgmm(model)
-      score_q = pryr::partial(scorefunctiongmm, model=model2)
-      result2 <- KSD(X,score_q=score_q, 'rbf',-1.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model2)
+      result2 <- KSD(X,score_function=score_function, 'rbf',-1.0)
       print(sprintf('Original p : %f',result$p))
       print(sprintf('Original p : %f',result2$p))
 
       print('3. 3-dimensional Gaussian, 100 samples, median-heuristic')
       model <- gmm(nComp=5,d=3)
       X <- rgmm(model, n=100)
-      score_q = pryr::partial(scorefunctiongmm, model=model)
-      result <- KSD(X,score_q=score_q, 'rbf',-2.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model)
+      result <- KSD(X,score_function=score_function, 'rbf',-2.0)
 
       model2 <- perturbgmm(model)
-      score_q = pryr::partial(scorefunctiongmm, model=model2)
-      result2 <- KSD(X,score_q=score_q, 'rbf',-2.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model2)
+      result2 <- KSD(X,score_function=score_function, 'rbf',-2.0)
       print(sprintf('Original p : %f',result$p))
       print(sprintf('Original p : %f',result2$p))
 
       print('4. 3-dimensional Gaussian, 300 samples, median-heuristic')
       model <- gmm(nComp=5,d=3)
       X <- rgmm(model, n=300)
-      score_q = pryr::partial(scorefunctiongmm, model=model)
-      result <- KSD(X,score_q=score_q, 'rbf',-2.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model)
+      result <- KSD(X,score_function=score_function, 'rbf',-2.0)
 
       model2 <- perturbgmm(model)
-      score_q = pryr::partial(scorefunctiongmm, model=model2)
-      result2 <- KSD(X,score_q=score_q, 'rbf',-2.0)
+      score_function = pryr::partial(scorefunctiongmm, model=model2)
+      result2 <- KSD(X,score_function=score_function, 'rbf',-2.0)
       print(sprintf('Original p : %f',result$p))
       print(sprintf('Original p : %f',result2$p))
 }