+++ NEWS.md
@@ -1,2 +1,7 @@
 ### KSD 0.1.0
 Initial submission to CRAN
+
+### KSD 1.0.0
+Name of score function parameter in KSD was changed from
+'score_q' to 'score_function'.
+Descriptions have been added.